requirements.txt
flask==2.2.3
werkzeug==2.2.3
flask-limiter==3.3.0
gunicorn==21.2.0
google-cloud-storage==2.9.0
google-cloud-firestore==2.11.0
google-cloud-logging==3.6.0
google-cloud-secret-manager==2.16.0
maxminddb==2.4.0
cryptography==40.0.2
